Improve feedback speed when saving/sending a draft
